requirements.txt
requests>=2.31.0
python-frontmatter>=1.0.0
pyyaml>=6.0
openai>=1.0.0
anthropic>=0.3.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0